+++ changelog.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [3.0.0] - 2022-04-16
+### Added
+* directive ChangeSizeDirective was added
+* services were injected to FirstComponent
+* class/token LocalStorageService was added
+* service ConfigOptionsService was added
+* service/token ConstantsService was added
+* service GeneratorService was added
+### Changed
+* CartService was modified
+* genId generator was added to service GeneratorService
+
 
 ## [2.0.0] - 2022-04-05
 ### Added